actionGrammar: move dfaBenchmark from spec to bench script
